fix(player interaction): handle object destruction
Interaction was based around ontrigger enter and exit functions. The exit function was not called if the object was inside the trigger when destroyed leading to undesired behavior.
Now it searches all the intractable within range and  selects the closest one, ensures deselection that can handle the destruction of the referenced item
Fixes #60